requirements.txt
Django==4.2
djangorestframework==3.14
django-crispy-forms==2.0
django-formtools==2.4
python-escpos==3.0
celery==5.3
redis==5.0
reportlab==4.0
openpyxl==3.1
channels==4.0
channels-redis==4.1
daphne==4.0
qrcode==7.4
Pillow==10.0
django-pwa==1.1
psycopg2-binary==2.9.7
django-cors-headers==4.3.1
pandas==2.0.3
numpy==1.24.3
crispy-bootstrap5==0.7